third_party/WebKit/LayoutTests/imported/wpt/html/semantics/embedded-content/the-object-element/historical-expected.txt

This is a testharness.js-based test.
FAIL object legacycaller should not be supported assert_equals: typeof expected "object" but got "function"
Harness: the test ran to completion.
